# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GADAVIRAL API
  description: |
    Ga & Dangme online social community API. One backend for Website, Android and Windows.
    Auth: JWT bearer access tokens (15 min) + rotating refresh tokens (30 days).
    Google sign-in uses genuine OpenID Connect ID tokens verified against Google JWKS.
    Authentication emails are sent from GADAVIRAL <admin@gadaviral.com>.
    Full endpoint reference: docs/API.md
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /posts:
    get: { tags: [Posts], summary: 'Feed (recent|following|recommended), paginated', security: [{ bearerAuth: [] }], responses: { '200': { description: OK } } }
    post: { tags: [Posts], summary: Create post (verified email; moderated), security: [{ bearerAuth: [] }], responses: { '201': { description: OK }, '403': { description: EMAIL_NOT_VERIFIED } } }
...
  '/posts/{id}/share':
    post: { tags: [Posts], summary: Share (real row; unique per target), security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '201': { description: OK } } }
  '/posts/{id}/vote':
    post: { tags: [Posts], summary: Poll vote, security: [{ bearerAuth: [] }], parameters: [{ '$ref': '#/components/parameters/Id' }], responses: { '201': { description: OK } } }
...
  '/users/{username}/follow':
    post: { tags: [Users], summary: Follow, security: [{ bearerAuth: [] }], parameters: [{ name: username, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    delete: { tags: [Users], summary: Unfollow, security: [{ bearerAuth: [] }], parameters: [{ name: username, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  '/users/{username}/block':
    post: { tags: [Users], summary: 'Block (+remove mutual follows)', security: [{ bearerAuth: [] }], parameters: [{ name: username, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /groups:
    get: { tags: [Community], summary: List/search groups, responses: { '200': { description: OK } } }
    post: { tags: [Community], summary: Create group, security: [{ bearerAuth: [] }], responses: { '201': { description: OK } } }
...
  /reports:
    post: { tags: [Safety], summary: Report any entity, security: [{ bearerAuth: [] }], responses: { '201': { description: OK } } }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }